src/knowledge_graph/knowledge_integration.py
import asyncio
import json
import logging
from typing import Any, Dict, List, Optional, Tuple, Union
from dataclasses import dataclass, field
from enum import Enum, auto

import networkx as nx
import pandas as pd
import torch
import transformers
from transformers import AutoModelForQuestionAnswering, AutoTokenizer

# Local imports
from .schema import Entity, EntityType, Relationship
from .advanced_embedding import MultiModalEmbeddingFinetuner

class KnowledgeSourceType(Enum):
    """
    Enumeration of knowledge source types for tracking and management.
    """
    ACADEMIC_PAPER = auto()
    TECHNICAL_DOCUMENT = auto()
    RESEARCH_DATABASE = auto()
    EXPERT_INTERVIEW = auto()
    WEB_SOURCE = auto()
    INTERNAL_REPOSITORY = auto()

@dataclass
class KnowledgeSource:
    """
    Comprehensive representation of a knowledge source.
    """
    source_type: KnowledgeSourceType
    identifier: str
    url: Optional[str] = None
    last_updated: Optional[str] = None
    reliability_score: float = 0.7
    metadata: Dict[str, Any] = field(default_factory=dict)

class KnowledgeGraphInterface:
    """
    Advanced knowledge graph management system.
    
    Key Capabilities:
    - Scalable knowledge representation
    - Multi-modal entity and relationship tracking
    - Dynamic knowledge integration
    """
    
    def __init__(
        self, 
        embedding_model: MultiModalEmbeddingFinetuner,
        max_entities: int = 100000
    ):
        """
        Initialize knowledge graph interface.
        
        Args:
            embedding_model: Multi-modal embedding fine-tuner
            max_entities: Maximum number of entities to store
        """
        self.graph = nx.DiGraph()
        self.embedding_model = embedding_model
        self.max_entities = max_entities
        
        # Knowledge source management
        self.knowledge_sources: Dict[str, KnowledgeSource] = {}
        
        # Logging configuration
        self.logger = logging.getLogger(__name__)
        
        # Caching mechanism for efficient retrieval
        self.entity_cache = {}
        self.relationship_cache = {}
    
    def add_entity(
        self, 
        entity: Entity, 
        source: Optional[KnowledgeSource] = None
    ) -> bool:
        """
        Add an entity to the knowledge graph.
        
        Args:
            entity: Entity to add
            source: Knowledge source of the entity
        
        Returns:
            Whether entity was successfully added
        """
        # Check graph size and manage capacity
        if len(self.graph.nodes) >= self.max_entities:
            self._prune_least_connected_entities()
        
        # Generate multi-modal embedding
        try:
            embedding = self.embedding_model.generate_entity_embedding(entity)
        except Exception as e:
            self.logger.error(f"Embedding generation failed: {e}")
            return False
        
        # Add node with embedding and metadata
        self.graph.add_node(
            entity.name, 
            type=entity.type, 
            embedding=embedding,
            source=source
        )
        
        # Update caches
        self.entity_cache[entity.name] = entity
        
        return True
    
    def add_relationship(
        self, 
        source_entity: Entity, 
        target_entity: Entity, 
        relationship_type: str
    ) -> bool:
        """
        Add a relationship between entities.
        
        Args:
            source_entity: Source entity
            target_entity: Target entity
            relationship_type: Type of relationship
        
        Returns:
            Whether relationship was successfully added
        """
        # Ensure both entities exist in graph
        if not (self.graph.has_node(source_entity.name) and 
                self.graph.has_node(target_entity.name)):
            return False
        
        # Add edge with relationship metadata
        self.graph.add_edge(
            source_entity.name, 
            target_entity.name, 
            type=relationship_type
        )
        
        # Update relationship cache
        cache_key = (source_entity.name, target_entity.name)
        self.relationship_cache[cache_key] = relationship_type
        
        return True
    
    def _prune_least_connected_entities(self):
        """
        Remove least connected entities when graph reaches capacity.
        """
        # Compute node connectivity
        connectivity = nx.degree_centrality(self.graph)
        
        # Sort nodes by connectivity
        sorted_nodes = sorted(
            connectivity.items(), 
            key=lambda x: x[1]
        )
        
        # Remove least connected nodes
        nodes_to_remove = [
            node for node, _ in sorted_nodes[:self.max_entities // 10]
        ]
        
        self.graph.remove_nodes_from(nodes_to_remove)
        
        # Clean up caches
        for node in nodes_to_remove:
            self.entity_cache.pop(node, None)

class ReasoningEngine:
    """
    Advanced reasoning engine for multi-hop reasoning and uncertainty management.
    
    Key Capabilities:
    - Multi-hop reasoning across knowledge graph
    - Uncertainty quantification
    - Contradiction resolution
    """
    
    def __init__(
        self, 
        knowledge_graph: KnowledgeGraphInterface,
        uncertainty_threshold: float = 0.3
    ):
        """
        Initialize reasoning engine.
        
        Args:
            knowledge_graph: Knowledge graph interface
            uncertainty_threshold: Threshold for uncertainty
        """
        self.knowledge_graph = knowledge_graph
        self.uncertainty_threshold = uncertainty_threshold
        
        # Language model for reasoning
        self.language_model = AutoModelForQuestionAnswering.from_pretrained(
            'deepset/roberta-base-squad2'
        )
        self.tokenizer = AutoTokenizer.from_pretrained(
            'deepset/roberta-base-squad2'
        )
    
    def multi_hop_reasoning(
        self, 
        query: str, 
        max_hops: int = 3
    ) -> Dict[str, Any]:
        """
        Perform multi-hop reasoning across knowledge graph.
        
        Args:
            query: Reasoning query
            max_hops: Maximum number of reasoning hops
        
        Returns:
            Reasoning results with confidence and explanation
        """
        # Initial entity identification
        initial_entities = self._identify_relevant_entities(query)
        
        reasoning_paths = []
        for entity in initial_entities:
            path = self._explore_reasoning_path(
                entity, 
                query, 
                max_hops=max_hops
            )
            reasoning_paths.append(path)
        
        # Aggregate and rank reasoning paths
        best_path = max(reasoning_paths, key=lambda p: p['confidence'])
        
        return best_path
    
    def _identify_relevant_entities(
        self, 
        query: str
    ) -> List[Entity]:
        """
        Identify entities relevant to the query.
        
        Args:
            query: Input query
        
        Returns:
            List of relevant entities
        """
        # Use language model for entity extraction
        inputs = self.tokenizer(
            query, 
            return_tensors='pt', 
            truncation=True, 
            max_length=512
        )
        
        # Extract potential entities
        outputs = self.language_model(**inputs)
        
        # TODO: Implement advanced entity extraction logic
        return []
    
    def _explore_reasoning_path(
        self, 
        start_entity: Entity, 
        query: str, 
        max_hops: int = 3
    ) -> Dict[str, Any]:
        """
        Explore reasoning path from a starting entity.
        
        Args:
            start_entity: Starting entity for reasoning
            query: Original reasoning query
            max_hops: Maximum number of reasoning hops
        
        Returns:
            Reasoning path with confidence and explanation
        """
        # Perform graph traversal
        reasoning_path = {
            'entities': [start_entity],
            'relationships': [],
            'confidence': 0.0,
            'explanation': ""
        }
        
        # TODO: Implement advanced path exploration logic
        
        return reasoning_path

class ResearchDirectionGenerator:
    """
    Advanced research direction generation system.
    
    Key Capabilities:
    - Research gap identification
    - Priority ranking
    - Experiment suggestion
    """
    
    def __init__(
        self, 
        knowledge_graph: KnowledgeGraphInterface,
        reasoning_engine: ReasoningEngine
    ):
        """
        Initialize research direction generator.
        
        Args:
            knowledge_graph: Knowledge graph interface
            reasoning_engine: Reasoning engine
        """
        self.knowledge_graph = knowledge_graph
        self.reasoning_engine = reasoning_engine
        
        # Research tracking
        self.research_gaps = {}
        self.research_priorities = {}
    
    def identify_research_gaps(
        self, 
        domain: str, 
        existing_research: List[str]
    ) -> Dict[str, float]:
        """
        Identify research gaps in a specific domain.
        
        Args:
            domain: Research domain
            existing_research: List of existing research topics
        
        Returns:
            Research gaps with priority scores
        """
        # Use reasoning engine to analyze knowledge gaps
        reasoning_query = f"What are the unexplored areas in {domain}?"
        reasoning_result = self.reasoning_engine.multi_hop_reasoning(reasoning_query)
        
        # Gap identification logic
        research_gaps = {}
        
        # TODO: Implement advanced gap identification
        
        return research_gaps
    
    def generate_research_roadmap(
        self, 
        initial_focus: str, 
        time_horizon: int = 5
    ) -> Dict[str, Any]:
        """
        Generate a technology research roadmap.
        
        Args:
            initial_focus: Starting research focus
            time_horizon: Research roadmap duration in years
        
        Returns:
            Detailed research roadmap
        """
        # Identify initial research priorities
        research_priorities = self.identify_research_gaps(
            initial_focus, 
            existing_research=[]
        )
        
        # Roadmap generation
        roadmap = {
            'initial_focus': initial_focus,
            'time_horizon': time_horizon,
            'research_stages': []
        }
        
        # TODO: Implement advanced roadmap generation
        
        return roadmap

def initialize_knowledge_integration_system(
    embedding_model: Optional[MultiModalEmbeddingFinetuner] = None
) -> Tuple[
    KnowledgeGraphInterface, 
    ReasoningEngine, 
    ResearchDirectionGenerator
]:
    """
    Initialize the complete knowledge integration system.
    
    Args:
        embedding_model: Optional multi-modal embedding fine-tuner
    
    Returns:
        Initialized knowledge integration components
    """
    # Create embedding model if not provided
    if embedding_model is None:
        embedding_model = MultiModalEmbeddingFinetuner()
    
    # Initialize knowledge graph
    knowledge_graph = KnowledgeGraphInterface(
        embedding_model=embedding_model
    )
    
    # Initialize reasoning engine
    reasoning_engine = ReasoningEngine(
        knowledge_graph=knowledge_graph
    )
    
    # Initialize research direction generator
    research_direction_generator = ResearchDirectionGenerator(
        knowledge_graph=knowledge_graph,
        reasoning_engine=reasoning_engine
    )
    
    return (
        knowledge_graph, 
        reasoning_engine, 
        research_direction_generator
    )
